# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Crypto Market Intelligence API
  version: 1.0.0
  description: Real-time crypto prices, alerts, and portfolio tracking API.
servers:
  - url: http://localhost:4000
    description: Development server

paths:
  /api/auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                name:
                  type: string
      responses:
        '201':
          description: Created
          
  /api/auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: OK
          
  /api/prices/live:
    get:
      summary: Get live prices
      responses:
        '200':
          description: OK
          
  /api/prices/history/{coinId}:
    get:
      summary: Get price history for a coin
      parameters:
        - in: path
          name: coinId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          
  /api/alerts:
    post:
      summary: Create a price alert
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                coinId:
                  type: string
                condition:
                  type: string
                targetPrice:
                  type: number
      responses:
        '201':
          description: Created
    get:
      summary: Get all active alerts for user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          
  /api/portfolio:
    get:
      summary: Get user portfolio and P&L
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          
  /api/portfolio/positions:
    post:
      summary: Add a portfolio position
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                coinId:
                  type: string
                quantity:
                  type: number
                purchasePrice:
                  type: number
      responses:
        '201':
          description: Created

  /api/admin/health:
    get:
      summary: Get system health
      responses:
        '200':
          description: OK

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT